enriched/cloud_function/requirements.txt
pandas
pyarrow
fsspec
gcsfs
scipy
google-cloud-storage
